requirements.txt
Flask==3.0.3
gunicorn==23.0.0
Werkzeug==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0
pytz==2024.1
SQLAlchemy~=2.0.40
pip~=23.2.1
wheel~=0.41.2
Jinja2~=3.1.6
setuptools~=68.2.0
typing_extensions~=4.13.1
packaging~=24.2
pytest
pytest-flask
Flask-Limiter==3.5.0